Burteforce transfers searching added, always runs out of time

+++ dicewars/ai/move.py
@@ -8,14 +8,14 @@ from dicewars.client.game.area import Area
 from dicewars.ai.aliases import Command, Name
 
 class Move:
-    def __init__(self, command: Command, scores: Optional[List[int]]=None, children: List['Move']=list(), target_name: Optional[int]=None):
+    def __init__(self, command: Command, scores: Optional[List[float]]=None, children: List['Move']=list(), target_name: Optional[int]=None):
         self.command = command
         self.scores = scores if scores else next.scores
         self.children = children
         self.target_name = target_name
 
     @classmethod
-    def other(cls, command, scores: Optional[List[int]]=None, next: Optional['Move']=None):
+    def other(cls, command, scores: Optional[List[float]]=None, next: Optional['Move']=None):
         return cls(command, scores if scores else next.scores, [next] if next else list(), None)
 
     @classmethod
